# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookCut API",
    "version": "1.0.0",
    "description": "API for managing barbershop appointments and services"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "paths": {
    "/booking/services": {
      "get": {
        "summary": "Get all services",
        "tags": ["Booking"],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/booking/barbers": {
      "get": {
        "summary": "Get all barbers",
        "tags": ["Booking"],
        "responses": {
          "200": {
            "description": "List of available barbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Barber"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/verify-code": {
      "post": {
        "summary": "Verify admin access code",
        "tags": ["Admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "ADMIN123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access granted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Access granted"
                    },
                    "token": {
                      "type": "string",
                      "example": "admin-token"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid access code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid access code"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/bookings": {
      "get": {
        "summary": "Get all bookings",
        "tags": ["Admin"],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/bookings/{id}/status": {
      "post": {
        "summary": "Update booking status",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "accepted", "declined"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/bookings/{id}/reschedule": {
      "post": {
        "summary": "Reschedule booking",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["appointment_date"],
                "properties": {
                  "appointment_date": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking rescheduled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/barbers": {
      "get": {
        "summary": "Get all barbers",
        "tags": ["Admin"],
        "responses": {
          "200": {
            "description": "List of all barbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Barber"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new barber",
        "tags": ["Admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Barber"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Barber created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Barber"
                }
              }
            }
          }
        }
      }
    },
    "/admin/barbers/{id}": {
      "get": {
        "summary": "Get barber by ID",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Barber details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Barber"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update barber",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Barber"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Barber updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Barber"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete barber",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Barber deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
      "Barber": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customer_name": {
            "type": "string"
          },
          "customer_email": {
            "type": "string"
          },
          "appointment_date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "accepted", "declined"]
          },
          "notes": {
            "type": "string"
          },
          "service_name": {
            "type": "string"
          },
          "service_duration": {
            "type": "integer"
          },
          "barber_name": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}